Split timezone utilities into separate files
