+++ packages/labs/src/components/timezone-picker/timezoneMetadata.ts
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ * Licensed under the BSD-3 License as modified (the “License”); you may obtain a copy
+ * of the license at https://github.com/palantir/blueprint/blob/master/LICENSE
+ * and https://github.com/palantir/blueprint/blob/master/PATENTS
+ */
+
+import * as moment from "moment-timezone";
+
+export interface ITimezoneMetadata {
+    timezone: string;
+    abbreviation: string;
+    offset: number;
+    offsetAsString: string;
+    population: number;
+}
+
+export function getTimezoneMetadata(timezone: string, date: Date): ITimezoneMetadata {
+    const timestamp = date.getTime();
+    const zone = moment.tz.zone(timezone);
+    const zonedDate = moment.tz(timestamp, timezone);
+    const offset = zonedDate.utcOffset();
+    const offsetAsString = zonedDate.format("Z");
+    const abbreviation = getAbbreviation(timezone, timestamp);
+    // TODO: amend moment-timezone typings to include the population field
+    const population = (zone && (zone as any).population) || 0;
+
+    return {
+        timezone,
+        abbreviation,
+        offset,
+        offsetAsString,
+        population,
+    };
+}
+
+function getAbbreviation(timezone: string, timestamp: number): string {
+    const zone = moment.tz.zone(timezone);
+    if (zone) {
+        const abbreviation = zone.abbr(timestamp);
+
+        // Only include abbreviations that are not just a repeat of the offset
+        if (abbreviation.length > 0 && abbreviation[0] !== "-" && abbreviation[0] !== "+") {
+            return abbreviation;
+        }
+    }
+
+    return "";
+}
