+++ packages/labs/src/components/timezone-picker/timezonePicker.tsx
@@ -6,7 +6,6 @@
  */
 
 import * as classNames from "classnames";
-import * as moment from "moment-timezone";
 import * as PureRender from "pure-render-decorator";
 import * as React from "react";
 
@@ -15,7 +14,6 @@ import {
     Button,
     Classes as CoreClasses,
     IButtonProps,
-    IconName,
     IPopoverProps,
     IProps,
     MenuItem,
@@ -26,6 +24,11 @@ import {
     Select,
 } from "..";
 import * as Classes from "../../common/classes";
+import { formatTimezone, TimezoneDisplayFormat } from "./timezoneDisplayFormat";
+import { getInitialTimezoneItems, getTimezoneItems, ITimezoneItem } from "./timezoneItems";
+import { getTimezoneQueryCandidates, isQueryMatch, ITimezoneQueryCandidates } from "./timezoneQueryUtils";
+
+export { TimezoneDisplayFormat };
 
 export interface ITimezonePickerProps extends IProps {
     /**
@@ -84,25 +87,10 @@ export interface ITimezonePickerProps extends IProps {
     popoverProps?: Partial<IPopoverProps> & object;
 }
 
-export type TimezoneDisplayFormat = "offset" | "abbreviation" | "name";
-export const TimezoneDisplayFormat = {
-    ABBREVIATION: "abbreviation" as "abbreviation",
-    NAME: "name" as "name",
-    OFFSET: "offset" as "offset",
-};
-
 export interface ITimezonePickerState {
     date?: Date;
     value?: string;
     query?: string;
-}
-
-interface ITimezoneItem {
-    key: string;
-    text: string;
-    label: string;
-    iconName?: IconName;
-    timezone: string;
 }
 
 const TypedSelect = Select.ofType<ITimezoneItem>();
@@ -118,8 +106,8 @@ export class TimezonePicker extends AbstractComponent<ITimezonePickerProps, ITim
         showLocalTimezone: true,
     };
 
-    private timezones: ITimezoneItem[];
-    private timezoneToQueryCandidates: { [timezone: string]: string[] };
+    private timezoneItems: ITimezoneItem[];
+    private timezoneToQueryCandidates: ITimezoneQueryCandidates;
     private initialTimezones: ITimezoneItem[];
 
     constructor(props: ITimezonePickerProps, context?: any) {
@@ -144,7 +132,7 @@ export class TimezonePicker extends AbstractComponent<ITimezonePickerProps, ITim
         return (
             <TypedSelect
                 className={classNames(Classes.TIMEZONE_PICKER, className)}
-                items={query ? this.timezones : this.initialTimezones}
+                items={query ? this.timezoneItems : this.initialTimezones}
                 itemListPredicate={this.filterTimezones}
                 itemRenderer={this.renderItem}
                 noResults={<MenuItem disabled text="No matching timezones." />}
@@ -205,18 +193,18 @@ export class TimezonePicker extends AbstractComponent<ITimezonePickerProps, ITim
     }
 
     private updateTimezones(date: Date): void {
-        const timezones = getTimezoneItems(date);
-        this.timezones = timezones;
+        const timezoneItems = getTimezoneItems(date);
+        const timezones = timezoneItems.map(({ timezone }) => timezone);
+        this.timezoneItems = timezoneItems;
         this.timezoneToQueryCandidates = getTimezoneQueryCandidates(timezones, date);
     }
 
     private filterTimezones = (query: string, items: ITimezoneItem[]): ITimezoneItem[] => {
-        const normalizedQuery = normalizeText(query);
         return items.filter((item) => {
             const candidates = this.timezoneToQueryCandidates[item.timezone] !== undefined
                 ? this.timezoneToQueryCandidates[item.timezone]
-                : [normalizeText(item.timezone)];
-            return candidates.some((candidate) => isQueryMatch(normalizedQuery, candidate));
+                : [item.timezone];
+            return candidates.some((candidate) => isQueryMatch(query, candidate));
         });
     }
 
@@ -250,193 +238,4 @@ export class TimezonePicker extends AbstractComponent<ITimezonePickerProps, ITim
     private handleQueryChange = (query: string) => {
         this.setState({ query });
     }
-}
-
-function getTimezoneItems(date: Date): ITimezoneItem[] {
-    return moment.tz.names().map((timezone) => toTimezoneItem(timezone, date));
-}
-
-function getInitialTimezoneItems(date: Date, showLocalTimezone: boolean): ITimezoneItem[] {
-    const populous = getPopulousTimezoneItems(date);
-    const local = getLocalTimezoneItem(date);
-    return showLocalTimezone && local !== undefined
-        ? [local, ...populous]
-        : populous;
-}
-
-function getPopulousTimezoneItems(date: Date): ITimezoneItem[] {
-    // Filter out noisy timezones. See https://github.com/moment/moment-timezone/issues/227
-    const timezones = moment.tz.names().filter((timezone) => /\//.test(timezone) && !/Etc\//.test(timezone));
-
-    const timezoneToMetadata: { [timezone: string]: ITimezoneMetadata } = {};
-    for (const timezone of timezones) {
-        timezoneToMetadata[timezone] = getTimezoneMetadata(timezone, date);
-    }
-
-    // Order by offset ascending, population descending, timezone name ascending
-    timezones.sort((timezone1, timezone2) => {
-        const { offset: offset1, population: population1 } = timezoneToMetadata[timezone1];
-        const { offset: offset2, population: population2 } = timezoneToMetadata[timezone2];
-        if (offset1 === offset2) {
-            if (population1 === population2) {
-                // Fall back to sorting alphabetically
-                return timezone1 < timezone2 ? -1 : 1;
-            }
-            // Descending order of population
-            return population2 - population1;
-        }
-        // Ascending order of offset
-        return offset1 - offset2;
-    });
-
-    // Choose the most populous locations for each offset
-    const initialTimezones: ITimezoneItem[] = [];
-    let prevOffset: number;
-    for (const timezone of timezones) {
-        const curOffset = timezoneToMetadata[timezone].offset;
-        if (prevOffset === undefined || prevOffset !== curOffset) {
-            initialTimezones.push(toTimezoneItem(timezone, date));
-            prevOffset = curOffset;
-        }
-    }
-    return initialTimezones;
-}
-
-function getLocalTimezoneItem(date: Date): ITimezoneItem | undefined {
-    const timezone = getLocalTimezone();
-    if (timezone !== undefined) {
-        const timestamp = date.getTime();
-        const zonedDate = moment.tz(timestamp, timezone);
-        const offsetAsString = zonedDate.format("Z");
-        return {
-            iconName: "locate",
-            key: `${timezone}-local`,
-            label: offsetAsString,
-            text: "Current timezone",
-            timezone,
-        };
-    } else {
-        return undefined;
-    }
-}
-
-/**
- * Get the local timezone.
- * Note that we are not guaranteed to get the correct timezone in all browsers,
- * so this is a best guess.
- * https://momentjs.com/timezone/docs/#/using-timezones/guessing-user-timezone/
- * https://github.com/moment/moment-timezone/blob/develop/moment-timezone.js#L328-L361
- */
-function getLocalTimezone(): string | undefined {
-    return moment.tz.guess();
-}
-
-function getTimezoneQueryCandidates(timezones: ITimezoneItem[], date: Date): { [timezone: string]: string[] } {
-    const timezoneToQueryCandidates: { [timezone: string]: string[] } = {};
-
-    for (const { timezone } of timezones) {
-        const { abbreviation, offsetAsString } = getTimezoneMetadata(timezone, date);
-        const candidates = [
-            timezone,
-            abbreviation,
-            offsetAsString,
-
-            // Split timezone string.
-            // For example, "America/Los_Angeles" -> "America Los Angeles"
-            timezone.split(/[/_]/).join(" "),
-
-            // Don't require leading offset 0.
-            // For example, "-07:00" -> "-7:00"
-            offsetAsString.replace(/^([+-])0(\d:\d{2})$/, "$1$2"),
-        ].map(normalizeText);
-
-        timezoneToQueryCandidates[timezone] = timezoneToQueryCandidates[timezone] || [];
-        timezoneToQueryCandidates[timezone].push(...candidates);
-    }
-
-    return timezoneToQueryCandidates;
-}
-
-function toTimezoneItem(timezone: string, date: Date): ITimezoneItem {
-    const { abbreviation, offsetAsString } = getTimezoneMetadata(timezone, date);
-    return {
-        key: timezone,
-        label: offsetAsString,
-        text: timezone + (abbreviation ? ` (${abbreviation})` : ""),
-        timezone,
-    };
-}
-
-interface ITimezoneMetadata {
-    timezone: string;
-    abbreviation: string;
-    offset: number;
-    offsetAsString: string;
-    population: number;
-}
-
-function getTimezoneMetadata(timezone: string, date: Date): ITimezoneMetadata {
-    const timestamp = date.getTime();
-    const zone = moment.tz.zone(timezone);
-    const zonedDate = moment.tz(timestamp, timezone);
-    const offset = zonedDate.utcOffset();
-    const offsetAsString = zonedDate.format("Z");
-    const abbreviation = getAbbreviation(timezone, timestamp);
-    const population = (zone && (zone as any).population) || 0;
-
-    return {
-        timezone,
-        abbreviation,
-        offset,
-        offsetAsString,
-        population,
-    };
-}
-
-function getAbbreviation(timezone: string, timestamp: number): string {
-    const zone = moment.tz.zone(timezone);
-    if (zone) {
-        const abbreviation = zone.abbr(timestamp);
-
-        // Only include abbreviations that are not just a repeat of the offset
-        if (abbreviation.length > 0 && abbreviation[0] !== "-" && abbreviation[0] !== "+") {
-            return abbreviation;
-        }
-    }
-
-    return "";
-}
-
-function formatTimezone(
-    timezone: string | undefined,
-    date: Date,
-    targetDisplayFormat: TimezoneDisplayFormat,
-): string | undefined {
-    if (!timezone || !moment.tz.zone(timezone)) {
-        return undefined;
-    }
-
-    switch (targetDisplayFormat) {
-        case TimezoneDisplayFormat.ABBREVIATION:
-            return moment.tz(date.getTime(), timezone).format("z");
-        case TimezoneDisplayFormat.NAME:
-            return timezone;
-        case TimezoneDisplayFormat.OFFSET:
-            return moment.tz(date.getTime(), timezone).format("Z");
-        default:
-            assertNever(targetDisplayFormat);
-            return "";
-    }
-}
-
-function isQueryMatch(query: string, candidate: string) {
-    return candidate.indexOf(query) >= 0;
-}
-
-function normalizeText(text: string): string {
-    return text.toLowerCase();
-}
-
-function assertNever(x: never): never {
-    throw new Error("Unexpected value: " + x);
 }
