+++ packages/labs/src/components/timezone-picker/timezoneUtils.ts
@@ -1,0 +1,19 @@
+/*
+ * Copyright 2017 Palantir Technologies, Inc. All rights reserved.
+ * Licensed under the BSD-3 License as modified (the “License”); you may obtain a copy
+ * of the license at https://github.com/palantir/blueprint/blob/master/LICENSE
+ * and https://github.com/palantir/blueprint/blob/master/PATENTS
+ */
+
+import * as moment from "moment-timezone";
+
+/**
+ * Get the user's local timezone.
+ * Note that we are not guaranteed to get the correct timezone in all browsers,
+ * so this is a best guess.
+ * https://momentjs.com/timezone/docs/#/using-timezones/guessing-user-timezone/
+ * https://github.com/moment/moment-timezone/blob/develop/moment-timezone.js#L328-L361
+ */
+export function getLocalTimezone(): string | undefined {
+    return moment.tz.guess();
+}
